refactor(slack): split normalize.ts shared helpers into domain modules
Second reorg step (per the Module Organization convention). Extract the
shared concerns out of the ~1,090-line normalize.ts into domain modules,
leaving only the normalizers behind for the next PR to split by family:

- message-schemas.ts — the tolerant Zod ingress schemas, their z.infer
  types (SlackFile, SlackMessageChangedEvent, SlackReactionEvent, …), the
  @slack/types cross-checks, and the shared output contract
  (NormalizedSlackEvent, SlackBotSenderInfo). This is the leaf module the
  others depend on.
- actor.ts — actor/bot-sender helpers: slackUserActorFields,
  slackBotSenderInfo, slackBotContactNote, enrichNormalizedActor.
- attachments.ts — extractSlackAttachments / extractSlackFileMap.
- render-text.ts — renderSlackInboundText + SlackTextRenderContext.

normalize.ts keeps buildNormalizedSlackMessage and the normalizers and
imports the above. Consumers (socket-mode, envelope, classify-event,
download, index, tests) repoint their type/helper imports to the new
modules; the normalizer functions still come from normalize.js.

Pure move — no logic change. 298 slack tests pass; tsc / eslint / knip
(no new unused exports) clean.
