+++ gateway/src/__tests__/slack-normalize.test.ts
@@ -4,11 +4,13 @@ import {
   normalizeSlackChannelMessage,
   normalizeSlackDirectMessage,
   normalizeSlackMessageEdit,
-  type SlackAppMentionEvent,
-  type SlackChannelMessageEvent,
-  type SlackDirectMessageEvent,
-  type SlackMessageChangedEvent,
 } from "../slack/normalize.js";
+import type {
+  SlackAppMentionEvent,
+  SlackChannelMessageEvent,
+  SlackDirectMessageEvent,
+  SlackMessageChangedEvent,
+} from "../slack/message-schemas.js";
 import type { GatewayConfig } from "../config.js";
 
 function makeConfig(overrides: Partial<GatewayConfig> = {}): GatewayConfig {
