+++ gateway/src/__tests__/slack-socket-mode-catchup.test.ts
@@ -4,7 +4,7 @@ import { drizzle } from "drizzle-orm/bun-sqlite";
 import type { GatewayConfig } from "../config.js";
 import { SlackStore } from "../db/slack-store.js";
 import * as schema from "../db/schema.js";
-import type { NormalizedSlackEvent } from "../slack/normalize.js";
+import type { NormalizedSlackEvent } from "../slack/message-schemas.js";
 
 type FetchFn = (
   input: string | URL | Request,
