+++ gateway/src/slack/download.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, mock, test } from "bun:test";
 import { ContentMismatchError } from "../download-validation.js";
-import type { SlackFile } from "./normalize.js";
+import type { SlackFile } from "./message-schemas.js";
 
 type FetchFn = (
   input: string | URL | Request,
@@ -142,9 +142,7 @@ describe("downloadSlackFile", () => {
   });
 
   test("throws when redirect has no Location header", async () => {
-    fetchMock = mock(
-      async () => new Response(null, { status: 302 }),
-    );
+    fetchMock = mock(async () => new Response(null, { status: 302 }));
 
     const file = makeSlackFile();
